Filter files by file extension
The engine currently does not expect include_paths to receive just `./`
causing the engine to not filter file extensions when `.codeclimate.yml`
doesn't have `exlude_paths`. This changes the entrypoint to always use
`include_paths` and work with `./`.

This updates the engine to have `DEFAULT_EXTENSIONS` as well as
configurable file extensions.