Change directory to `/code`
The engine currently receives `["./"]` as the default `include_paths`
which breaks the current implementation of `queueWithIncludePaths` since
it generates the invalid path `/code/./`.

Instead of building the path to `/code` each time we change the working
directory to be `/code` so we can use relative paths. This means `./`
points to `/code` instead of generating the invalid path.